+++ decidim-core/lib/decidim/content_parsers/base_parser.rb
@@ -5,7 +5,7 @@ module Decidim
     # Abstract base class for content parsers, so they have the same contract
     #
     # @example How to use a content parser class
-    #   parser = Decidim::ContentParsers::CustomParser.new(content)
+    #   parser = Decidim::ContentParsers::CustomParser.new(content, context)
     #   parser.rewrite # returns the content rewritten
     #   parser.metadata # returns a Metadata object
     #
@@ -17,11 +17,16 @@ module Decidim
       # @return [String] the content to be rewritten
       attr_reader :content
 
+      # @return [Hash] with context information
+      attr_reader :context
+
       # Gets initialized with the `content` to parse
       #
       # @param content [String] already rewritten content or regular content
-      def initialize(content)
+      # @param context [Hash] arbitrary information to have a context
+      def initialize(content, context)
         @content = content || ""
+        @context = context
       end
 
       # Parse the `content` and return it modified